# src/pkg/domain/model/InventoryMeta.pb.go
// Code generated by protoc-gen-go. DO NOT EDIT.
// versions:
// 	protoc-gen-go v1.34.1
// 	protoc        v4.25.1
// source: InventoryMeta.proto

package domain

import (
	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
	reflect "reflect"
	sync "sync"
)

const (
	// Verify that this generated code is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
	// Verify that runtime/protoimpl is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
)

type InventoryMeta struct {
	state         protoimpl.MessageState
	sizeCache     protoimpl.SizeCache
	unknownFields protoimpl.UnknownFields

	Id        int64  `protobuf:"varint,1,opt,name=id,proto3" json:"id,omitempty"`
	Name      string `protobuf:"bytes,2,opt,name=name,proto3" json:"name,omitempty"`
	UserId    int64  `protobuf:"varint,3,opt,name=userId,proto3" json:"userId,omitempty"`
	Width     int64  `protobuf:"varint,4,opt,name=width,proto3" json:"width,omitempty"`
	Height    int64  `protobuf:"varint,5,opt,name=height,proto3" json:"height,omitempty"`
	MaxWeight int64  `protobuf:"varint,6,opt,name=maxWeight,proto3" json:"maxWeight,omitempty"`
}

func (x *InventoryMeta) Reset() {
	*x = InventoryMeta{}
	if protoimpl.UnsafeEnabled {
		mi := &file_InventoryMeta_proto_msgTypes[0]
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		ms.StoreMessageInfo(mi)
	}
}

func (x *InventoryMeta) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*InventoryMeta) ProtoMessage() {}

func (x *InventoryMeta) ProtoReflect() protoreflect.Message {
	mi := &file_InventoryMeta_proto_msgTypes[0]
	if protoimpl.UnsafeEnabled && x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use InventoryMeta.ProtoReflect.Descriptor instead.
func (*InventoryMeta) Descriptor() ([]byte, []int) {
	return file_InventoryMeta_proto_rawDescGZIP(), []int{0}
}

func (x *InventoryMeta) GetId() int64 {
	if x != nil {
		return x.Id
	}
	return 0
}

func (x *InventoryMeta) GetName() string {
	if x != nil {
		return x.Name
	}
	return ""
}

func (x *InventoryMeta) GetUserId() int64 {
	if x != nil {
		return x.UserId
	}
	return 0
}

func (x *InventoryMeta) GetWidth() int64 {
	if x != nil {
		return x.Width
	}
	return 0
}

func (x *InventoryMeta) GetHeight() int64 {
	if x != nil {
		return x.Height
	}
	return 0
}

func (x *InventoryMeta) GetMaxWeight() int64 {
	if x != nil {
		return x.MaxWeight
	}
	return 0
}

var File_InventoryMeta_proto protoreflect.FileDescriptor

var file_InventoryMeta_proto_rawDesc = []byte{
	0x0a, 0x13, 0x49, 0x6e, 0x76, 0x65, 0x6e, 0x74, 0x6f, 0x72, 0x79, 0x4d, 0x65, 0x74, 0x61, 0x2e,
	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x06, 0x64, 0x6f, 0x6d, 0x61, 0x69, 0x6e, 0x22, 0x97, 0x01,
	0x0a, 0x0d, 0x49, 0x6e, 0x76, 0x65, 0x6e, 0x74, 0x6f, 0x72, 0x79, 0x4d, 0x65, 0x74, 0x61, 0x12,
	0x0e, 0x0a, 0x02, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x03, 0x52, 0x02, 0x69, 0x64, 0x12,
	0x12, 0x0a, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x6e,
	0x61, 0x6d, 0x65, 0x12, 0x16, 0x0a, 0x06, 0x75, 0x73, 0x65, 0x72, 0x49, 0x64, 0x18, 0x03, 0x20,
	0x01, 0x28, 0x03, 0x52, 0x06, 0x75, 0x73, 0x65, 0x72, 0x49, 0x64, 0x12, 0x14, 0x0a, 0x05, 0x77,
	0x69, 0x64, 0x74, 0x68, 0x18, 0x04, 0x20, 0x01, 0x28, 0x03, 0x52, 0x05, 0x77, 0x69, 0x64, 0x74,
	0x68, 0x12, 0x16, 0x0a, 0x06, 0x68, 0x65, 0x69, 0x67, 0x68, 0x74, 0x18, 0x05, 0x20, 0x01, 0x28,
	0x03, 0x52, 0x06, 0x68, 0x65, 0x69, 0x67, 0x68, 0x74, 0x12, 0x1c, 0x0a, 0x09, 0x6d, 0x61, 0x78,
	0x57, 0x65, 0x69, 0x67, 0x68, 0x74, 0x18, 0x06, 0x20, 0x01, 0x28, 0x03, 0x52, 0x09, 0x6d, 0x61,
	0x78, 0x57, 0x65, 0x69, 0x67, 0x68, 0x74, 0x42, 0x0a, 0x5a, 0x08, 0x2e, 0x2f, 0x64, 0x6f, 0x6d,
	0x61, 0x69, 0x6e, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
}

var (
	file_InventoryMeta_proto_rawDescOnce sync.Once
	file_InventoryMeta_proto_rawDescData = file_InventoryMeta_proto_rawDesc
)

func file_InventoryMeta_proto_rawDescGZIP() []byte {
	file_InventoryMeta_proto_rawDescOnce.Do(func() {
		file_InventoryMeta_proto_rawDescData = protoimpl.X.CompressGZIP(file_InventoryMeta_proto_rawDescData)
	})
	return file_InventoryMeta_proto_rawDescData
}

var file_InventoryMeta_proto_msgTypes = make([]protoimpl.MessageInfo, 1)
var file_InventoryMeta_proto_goTypes = []interface{}{
	(*InventoryMeta)(nil), // 0: domain.InventoryMeta
}
var file_InventoryMeta_proto_depIdxs = []int32{
	0, // [0:0] is the sub-list for method output_type
	0, // [0:0] is the sub-list for method input_type
	0, // [0:0] is the sub-list for extension type_name
	0, // [0:0] is the sub-list for extension extendee
	0, // [0:0] is the sub-list for field type_name
}

func init() { file_InventoryMeta_proto_init() }
func file_InventoryMeta_proto_init() {
	if File_InventoryMeta_proto != nil {
		return
	}
	if !protoimpl.UnsafeEnabled {
		file_InventoryMeta_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
			switch v := v.(*InventoryMeta); i {
			case 0:
				return &v.state
			case 1:
				return &v.sizeCache
			case 2:
				return &v.unknownFields
			default:
				return nil
			}
		}
	}
	type x struct{}
	out := protoimpl.TypeBuilder{
		File: protoimpl.DescBuilder{
			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
			RawDescriptor: file_InventoryMeta_proto_rawDesc,
			NumEnums:      0,
			NumMessages:   1,
			NumExtensions: 0,
			NumServices:   0,
		},
		GoTypes:           file_InventoryMeta_proto_goTypes,
		DependencyIndexes: file_InventoryMeta_proto_depIdxs,
		MessageInfos:      file_InventoryMeta_proto_msgTypes,
	}.Build()
	File_InventoryMeta_proto = out.File
	file_InventoryMeta_proto_rawDesc = nil
	file_InventoryMeta_proto_goTypes = nil
	file_InventoryMeta_proto_depIdxs = nil
}