# src/pkg/server/handler/apihandler/embedded/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Inventory Manager",
    "description": "Inventory Manager API",
    "version": "1.0.0",
    "contact": {
      "name": "LinuxCode",
      "email": "jbensch@linuxcode.net"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "inventory.linuxcode.net",
      "description": "Inventory server for example"
    }
  ],
...
  "paths": {
...
    "/api/users": {
      "get": {
        "summary": "Get all users",
        "description": "Retrieve all users",
        "operationId": "getAllUsers",
        "tags": [
          "user"
        ],
        "responses": {
          "200": {
            "description": "successful operation - returns a list of users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponse"
                },
                "example": {
                  "users": [
                    {
                      "id": 1,
                      "username": "john_doe",
                      "inventories": [
                        {
                          "id": 1,
                          "name": "Backpack",
                          "userId": 1,
                          "max_weight": 850,
                          "volume": {
                            "height": 3,
                            "width": 4
                          }
                        },
                        {
                          "id": 2,
                          "name": "Bank",
                          "userId": 1,
                          "max_weight": 10000,
                          "volume": {
                            "height": 10,
                            "width": 10
                          }
                        }
                      ]
                    },
                    {
                      "id": 2,
                      "username": "jane_doe",
                      "inventories": [
                        {
                          "id": 3,
                          "name": "Big Backpack",
                          "userId": 2,
                          "max_weight": 1350,
                          "volume": {
                            "height": 5,
                            "width": 8
                          }
                        },
                        {
                          "id": 4,
                          "name": "Small Bank",
                          "userId": 2,
                          "max_weight": 1000,
                          "volume": {
                            "height": 7,
                            "width": 7
                          }
                        }
                      ]
                    }
                  ],
                  "pagination": {
                    "page": 1,
                    "pages": 10,
                    "limit": 10,
                    "total": 100,
                    "offset": 0
                  }
                }
              }
            }
          },
          "404": {
            "description": "No users found - no users in the system",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "errorId": "123e4567-e89b-12d3-a456-426614174000",
                  "error": "The item does not exist",
                  "ErrorType": "Not Found",
                  "code": 404,
                  "timestamp": "2020-01-01T00:00:00Z"
                }
              }
            }
          }
        },
        "security": []
      },
      "post": {
        "summary": "Add a new user",
        "description": "Add a new user to the system",
        "operationId": "addUser",
        "tags": [
          "user"
        ],
        "requestBody": {
          "description": "User to add to the system",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPostRequest"
              },
              "example": {
                "username": "john_doe",
                "email": "something@something.com",
                "password": "password"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Added user - returns the newly created user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                },
                "example": {
                  "id": 1,
                  "username": "john_doe",
                  "inventories": [
                    {
                      "id": 1,
                      "name": "Backpack",
                      "userId": 1,
                      "max_weight": 850,
                      "volume": {
                        "height": 3,
                        "width": 4
                      }
                    },
                    {
                      "id": 2,
                      "name": "Bank",
                      "userId": 1,
                      "max_weight": 10000,
                      "volume": {
                        "height": 10,
                        "width": 10
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request - e.g. invalid user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "errorId": "123e4567-e89b-12d3-a456-426614174000",
                  "error": "The item does not exist",
                  "ErrorType": "Not Found",
                  "code": 404,
                  "timestamp": "2020-01-01T00:00:00Z"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/users/{userId}": {
      "get": {
        "summary": "Get a user",
        "description": "Get a user by ID",
        "operationId": "getUserById",
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/User"
                  },
                  "example": {
                    "id": 1,
                    "username": "john_doe",
                    "inventories": [
                      {
                        "id": 1,
                        "name": "Backpack",
                        "userId": 1,
                        "max_weight": 850,
                        "volume": {
                          "height": 3,
                          "width": 4
                        }
                      },
                      {
                        "id": 2,
                        "name": "Bank",
                        "userId": 1,
                        "max_weight": 10000,
                        "volume": {
                          "height": 10,
                          "width": 10
                        }
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied - wrong format of user ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "errorId": "123e4567-e89b-12d3-a456-426614174000",
                  "error": "The item does not exist",
                  "ErrorType": "Not Found",
                  "code": 404,
                  "timestamp": "2020-01-01T00:00:00Z"
                }
              }
            }
          },
          "404": {
            "description": "Can not find user (user ID) to return"
          }
        },
        "security": []
      },
      "put": {
        "summary": "Update a user",
        "description": "Update a user by ID",
        "operationId": "updateUserById",
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user that needs to be updated",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "required": true,
          "description": "User object that needs to be updated",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPostRequest"
              },
              "example": {
                "username": "john_doe",
                "email": "something@something.com",
                "password": "password"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update done - returns updated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                },
                "example": {
                  "id": 1,
                  "username": "john_doe",
                  "inventories": [
                    {
                      "id": 1,
                      "name": "Backpack",
                      "userId": 1,
                      "max_weight": 850,
                      "volume": {
                        "height": 3,
                        "width": 4
                      }
                    },
                    {
                      "id": 2,
                      "name": "Bank",
                      "userId": 1,
                      "max_weight": 10000,
                      "volume": {
                        "height": 10,
                        "width": 10
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request - wrong format of user ID or wrong format of user object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "errorId": "123e4567-e89b-12d3-a456-426614174000",
                  "error": "The item does not exist",
                  "ErrorType": "Not Found",
                  "code": 404,
                  "timestamp": "2020-01-01T00:00:00Z"
                }
              }
            }
          },
          "404": {
            "description": "Can not find user (user ID) to update"
          }
        },
        "security": []
      },
      "delete": {
        "summary": "Delete a user",
        "description": "Delete a user by ID",
        "operationId": "deleteUserById",
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted user (user ID)"
          },
          "400": {
            "description": "Invalid ID supplied - wrong format of user ID, deletion not possible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "errorId": "123e4567-e89b-12d3-a456-426614174000",
                  "error": "The item does not exist",
                  "ErrorType": "Not Found",
                  "code": 404,
                  "timestamp": "2020-01-01T00:00:00Z"
                }
              }
            }
          },
          "404": {
            "description": "Can not find user (user ID) to delete"
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Pagination": {
        "type": "object",
        "description": "This object holds the pagination data.",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "pages": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "limit": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "example": 100
          },
          "offset": {
            "type": "integer",
            "format": "int64",
            "example": 0
          }
        },
        "example": {
          "page": 1,
          "pages": 10,
          "limit": 10,
          "total": 100,
          "offset": 0
        },
        "required": [
          "page",
          "pages",
          "limit",
          "total",
          "offset"
        ]
      },
...
      "ErrorType": {
        "type": "string",
        "description": "The error type",
        "enum": [
          "Bad Request",
          "Not Found",
          "Not Implemented",
          "Internal Server Error"
        ],
        "example": "Bad Request"
      },
      "Error": {
        "type": "object",
        "description": "This object holds the error response data.",
        "properties": {
          "errorId": {
            "type": "string",
            "description": "The unique identifier for the error",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "error": {
            "type": "string",
            "description": "The error message",
            "example": "The item does not exist"
          },
          "ErrorType": {
            "$ref": "#/components/schemas/ErrorType"
          },
          "code": {
            "type": "integer",
            "description": "The error code",
            "example": 404
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "The time the error occurred",
            "example": "2020-01-01T00:00:00Z"
          }
        },
        "example": {
          "errorId": "123e4567-e89b-12d3-a456-426614174000",
          "error": "The item does not exist",
          "ErrorType": "Not Found",
          "code": 404,
          "timestamp": "2020-01-01T00:00:00Z"
        },
        "required": [
          "errorId",
          "error",
          "ErrorType",
          "code",
          "timestamp"
        ]
      },
      "Volume": {
        "type": "object",
        "description": "This object holds the volume data.",
        "properties": {
          "height": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "width": {
            "type": "integer",
            "format": "int64",
            "example": 10
          }
        },
        "example": {
          "height": 10,
          "width": 10
        },
        "required": [
          "height",
          "width"
        ]
      },
      "InventoryMeta": {
        "type": "object",
        "description": "This object holds the inventory meta data (restrictions, owner, etc.)",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Backpack"
          },
          "userId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "max_weight": {
            "type": "integer",
            "format": "int64",
            "example": 850
          },
          "volume": {
            "$ref": "#/components/schemas/Volume"
          }
        },
        "example": {
          "id": 1,
          "name": "Backpack",
          "userId": 1,
          "max_weight": 850,
          "volume": {
            "height": 3,
            "width": 4
          }
        },
        "required": [
          "id",
          "name",
          "userId",
          "max_weight",
          "volume"
        ]
      },
      "InventoryMetaList": {
        "type": "array",
        "description": "The list of inventory metas.",
        "items": {
          "$ref": "#/components/schemas/InventoryMeta"
        },
        "example": [
          {
            "id": 1,
            "name": "Backpack",
            "userId": 1,
            "max_weight": 850,
            "volume": {
              "height": 3,
              "width": 4
            }
          },
          {
            "id": 2,
            "name": "Chest",
            "userId": 1,
            "max_weight": 1500,
            "volume": {
              "height": 10,
              "width": 10
            }
          }
        ]
      },
      "User": {
        "type": "object",
        "description": "This object holds the user data.",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "username": {
            "type": "string",
            "example": "JohnDoe"
          },
          "inventories": {
            "$ref": "#/components/schemas/InventoryMetaList"
          }
        },
        "example": {
          "id": 1,
          "username": "john_doe",
          "inventories": [
            {
              "id": 1,
              "name": "Backpack",
              "userId": 1,
              "max_weight": 850,
              "volume": {
                "height": 3,
                "width": 4
              }
            },
            {
              "id": 2,
              "name": "Bank",
              "userId": 1,
              "max_weight": 10000,
              "volume": {
                "height": 10,
                "width": 10
              }
            }
          ]
        },
        "required": [
          "id",
          "username",
          "inventories"
        ]
      },
      "UserList": {
        "type": "array",
        "description": "The list of users.",
        "items": {
          "$ref": "#/components/schemas/User"
        },
        "example": [
          {
            "id": 1,
            "username": "john_doe",
            "inventories": [
              {
                "id": 1,
                "name": "Backpack",
                "userId": 1,
                "max_weight": 850,
                "volume": {
                  "height": 3,
                  "width": 4
                }
              },
              {
                "id": 2,
                "name": "Bank",
                "userId": 1,
                "max_weight": 10000,
                "volume": {
                  "height": 10,
                  "width": 10
                }
              }
            ]
          },
          {
            "id": 2,
            "username": "jane_doe",
            "inventories": [
              {
                "id": 3,
                "name": "Big Backpack",
                "userId": 2,
                "max_weight": 1350,
                "volume": {
                  "height": 5,
                  "width": 8
                }
              },
              {
                "id": 4,
                "name": "Small Bank",
                "userId": 2,
                "max_weight": 1000,
                "volume": {
                  "height": 7,
                  "width": 7
                }
              }
            ]
          }
        ]
      },
      "UserListResponse": {
        "type": "object",
        "description": "This object holds the user list response data. It includes the list of users and pagination data.",
        "properties": {
          "users": {
            "$ref": "#/components/schemas/UserList"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "example": {
          "users": [
            {
              "id": 1,
              "username": "john_doe",
              "inventories": [
                {
                  "id": 1,
                  "name": "Backpack",
                  "userId": 1,
                  "max_weight": 850,
                  "volume": {
                    "height": 3,
                    "width": 4
                  }
                },
                {
                  "id": 2,
                  "name": "Bank",
                  "userId": 1,
                  "max_weight": 10000,
                  "volume": {
                    "height": 10,
                    "width": 10
                  }
                }
              ]
            },
            {
              "id": 2,
              "username": "jane_doe",
              "inventories": [
                {
                  "id": 3,
                  "name": "Big Backpack",
                  "userId": 2,
                  "max_weight": 1350,
                  "volume": {
                    "height": 5,
                    "width": 8
                  }
                },
                {
                  "id": 4,
                  "name": "Small Bank",
                  "userId": 2,
                  "max_weight": 1000,
                  "volume": {
                    "height": 7,
                    "width": 7
                  }
                }
              ]
            }
          ],
          "pagination": {
            "page": 1,
            "pages": 10,
            "limit": 10,
            "total": 100,
            "offset": 0
          }
        },
        "required": [
          "users",
          "pagination"
        ]
      },
      "UserPostRequest": {
        "type": "object",
        "description": "This object holds the request data for creating a new user.",
        "properties": {
          "username": {
            "type": "string",
            "example": "john_doe"
          },
          "email": {
            "type": "string",
            "example": "john_doe@test.com"
          },
          "password": {
            "type": "string",
            "example": "password"
          }
        },
        "example": {
          "username": "john_doe",
          "email": "something@something.com",
          "password": "password"
        },
        "required": [
          "username",
          "email",
          "password"
        ]
      },
...
    }
  }
}